+++ test/base/action.js
@@ -6,7 +6,7 @@ var utils = require("../utils/test-utils")
 test("action should wrap in transaction", () => {
     var values = []
 
-    var observable = mobx.observable(0)
+    var observable = mobx.observable.box(0)
     var d = mobx.autorun(() => values.push(observable.get()))
 
     var increment = mobx.action("increment", amount => {
@@ -23,7 +23,7 @@ test("action should wrap in transaction", () => {
 })
 
 test("action modifications should be picked up 1", () => {
-    var a = mobx.observable(1)
+    var a = mobx.observable.box(1)
     var i = 3
     var b = 0
 
@@ -45,7 +45,7 @@ test("action modifications should be picked up 1", () => {
 })
 
 test("action modifications should be picked up 1", () => {
-    var a = mobx.observable(1)
+    var a = mobx.observable.box(1)
     var b = 0
 
     mobx.autorun(() => {
@@ -66,7 +66,7 @@ test("action modifications should be picked up 1", () => {
 })
 
 test("action modifications should be picked up 3", () => {
-    var a = mobx.observable(1)
+    var a = mobx.observable.box(1)
     var b = 0
 
     var doubler = mobx.computed(() => a.get() * 2)
@@ -89,8 +89,8 @@ test("action modifications should be picked up 3", () => {
 })
 
 test("test action should be untracked", () => {
-    var a = mobx.observable(3)
-    var b = mobx.observable(4)
+    var a = mobx.observable.box(3)
+    var b = mobx.observable.box(4)
     var latest = 0
     var runs = 0
 
@@ -131,7 +131,7 @@ test("test action should be untracked", () => {
 })
 
 test("should be possible to create autorun in ation", () => {
-    var a = mobx.observable(1)
+    var a = mobx.observable.box(1)
     var values = []
 
     var adder = mobx.action(inc => {
@@ -153,7 +153,7 @@ test("should be possible to create autorun in ation", () => {
 })
 
 test("should be possible to change unobserved state in an action called from computed", () => {
-    var a = mobx.observable(2)
+    var a = mobx.observable.box(2)
 
     var testAction = mobx.action(() => {
         a.set(3)
@@ -174,7 +174,7 @@ test("should be possible to change unobserved state in an action called from com
 })
 
 test("should not be possible to change observed state in an action called from computed", () => {
-    var a = mobx.observable(2)
+    var a = mobx.observable.box(2)
     var d = mobx.autorun(() => {
         a.get()
     })
@@ -199,8 +199,8 @@ test("should not be possible to change observed state in an action called from c
 })
 
 test("action in autorun should be untracked", () => {
-    var a = mobx.observable(2)
-    var b = mobx.observable(3)
+    var a = mobx.observable.box(2)
+    var b = mobx.observable.box(3)
 
     var data = []
     var multiplier = mobx.action(val => val * b.get())
@@ -284,7 +284,7 @@ test("runInAction", () => {
             })
     })
 
-    var observable = mobx.observable(0)
+    var observable = mobx.observable.box(0)
     var d = mobx.autorun(() => values.push(observable.get()))
 
     var res = mobx.runInAction("increment", () => {
@@ -348,7 +348,7 @@ test("action in autorun does not keep / make computed values alive", () => {
 test("computed values and actions", () => {
     let calls = 0
 
-    const number = mobx.observable(1)
+    const number = mobx.observable.box(1)
     const squared = mobx.computed(() => {
         calls++
         return number.get() * number.get()